requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
pyarrow>=16.0.0
fastparquet>=2024.2.0
requests>=2.32.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
duckdb>=0.10.0
pdfminer.six>=20221105
plotly>=5.22.0
openpyxl>=3.1.0
geopandas>=1.0.0
scikit-learn>=1.3.0
